+++ gensim/corpora/wikicorpus.py
@@ -12,11 +12,8 @@ Uses multiprocessing internally to parallelize the work and process the dump mor
 
 Notes
 -----
-If you have the `pattern <https://github.com/clips/pattern>`_ package installed,
-this module will use a fancy lemmatization to get a lemma of each token (instead of plain alphabetic tokenizer).
 
 See :mod:`gensim.scripts.make_wiki` for a canned (example) command-line script based on this module.
-
 """
 
 import bz2
@@ -467,9 +464,8 @@ def process_article(args, tokenizer_func=tokenize, token_min_len=TOKEN_MIN_LEN,
 
     Parameters
     ----------
-    args : (str, bool, str, int)
-        Article text, lemmatize flag (if True, :func:`~gensim.utils.lemmatize` will be used), article title,
-        page identificator.
+    args : (str, str, int)
+        Article text, article title, page identificator.
     tokenizer_func : function
         Function for tokenization (defaults is :func:`~gensim.corpora.wikicorpus.tokenize`).
         Needs to have interface:
@@ -487,12 +483,9 @@ def process_article(args, tokenizer_func=tokenize, token_min_len=TOKEN_MIN_LEN,
         List of tokens from article, title and page id.
 
     """
-    text, lemmatize, title, pageid = args
+    text, title, pageid = args
     text = filter_wiki(text)
-    if lemmatize:
-        result = utils.lemmatize(text)
-    else:
-        result = tokenizer_func(text, token_min_len, token_max_len, lower)
+    result = tokenizer_func(text, token_min_len, token_max_len, lower)
     return result, title, pageid
 
 
@@ -574,7 +567,7 @@ class WikiCorpus(TextCorpus):
         >>> MmCorpus.serialize(corpus_path, wiki)  # another 8h, creates a file in MatrixMarket format and mapping
 
     """
-    def __init__(self, fname, processes=None, lemmatize=utils.has_pattern(), dictionary=None,
+    def __init__(self, fname, processes=None, lemmatize=None, dictionary=None,
                  filter_namespaces=('0',), tokenizer_func=tokenize, article_min_tokens=ARTICLE_MIN_WORDS,
                  token_min_len=TOKEN_MIN_LEN, token_max_len=TOKEN_MAX_LEN, lower=True, filter_articles=None):
         """Initialize the corpus.
@@ -588,9 +581,6 @@ class WikiCorpus(TextCorpus):
             Path to the Wikipedia dump file.
         processes : int, optional
             Number of processes to run, defaults to `max(1, number of cpu - 1)`.
-        lemmatize : bool
-            Use lemmatization instead of simple regexp tokenization.
-            Defaults to `True` if you have the `pattern <https://github.com/clips/pattern>`_ package installed.
         dictionary : :class:`~gensim.corpora.dictionary.Dictionary`, optional
             Dictionary, if not provided,  this scans the corpus once, to determine its vocabulary
             **IMPORTANT: this needs a really long time**.
@@ -618,6 +608,13 @@ class WikiCorpus(TextCorpus):
         Unless a dictionary is provided, this scans the corpus once, to determine its vocabulary.
 
         """
+        if lemmatize is not None:
+            raise NotImplementedError(
+                'The lemmatize parameter is no longer supported. '
+                'If you need to lemmatize, use e.g. <https://github.com/clips/pattern>. '
+                'Perform lemmatization as part of your tokenization function and '
+                'pass it as the tokenizer_func parameter to this initializer.'
+            )
         self.fname = fname
         self.filter_namespaces = filter_namespaces
         self.filter_articles = filter_articles
@@ -625,7 +622,6 @@ class WikiCorpus(TextCorpus):
         if processes is None:
             processes = max(1, multiprocessing.cpu_count() - 1)
         self.processes = processes
-        self.lemmatize = lemmatize
         self.tokenizer_func = tokenizer_func
         self.article_min_tokens = article_min_tokens
         self.token_min_len = token_min_len
@@ -677,7 +673,7 @@ class WikiCorpus(TextCorpus):
 
         tokenization_params = (self.tokenizer_func, self.token_min_len, self.token_max_len, self.lower)
         texts = (
-            (text, self.lemmatize, title, pageid, tokenization_params)
+            (text, title, pageid, tokenization_params)
             for title, text, pageid
             in extract_pages(bz2.BZ2File(self.fname), self.filter_namespaces, self.filter_articles)
         )